Backend/pyproject.toml
[project]
name = "citysewa"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.36",
    "django>=5.0,<6.0",
    "django-cors-headers>=4.9.0",
    "django-environ>=0.12.0",
    "django-redis>=6.0.0",
    "djangorestframework==3.16",
    "drf-spectacular>=0.29.0",
    "gunicorn==23.0.0",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "sentry-sdk[django]>=2.52.0",
]
